Fix setup-owned IDs, lock span, and empty-engine load.
Setup IDs now use stable_id over organization and identity so hyphenated
pairs cannot collide and long identifiers stay bounded. Document mutation
copies unrelated top-level fields. Compensation holds one non-reentrant
lock across snapshot, selected write, the setup window, and restore.
Disabling through a symlink writes an empty valid document, and a file
with no policies loads as the unguarded hot path.
